Chop tree
Have done with find and chop tree without animation

+++ Assets/_OurData/_Scripts/Res/Resource.cs
@@ -7,5 +7,5 @@ using UnityEngine;
 public class Resource 
 {
     public ResourceName name;
-    public int number;
+    public float number;
 }
